@@ -8,24 +8,38 @@
     "build": "tsc",
     "start": "node dist/index.js",
     "migrate": "prisma migrate dev --name init",
-    "generate": "prisma generate"
+    "generate": "prisma generate",
+    "test": "jest"
   },
   "dependencies": {
-    "@prisma/client": "^4.0.0",
+    "@prisma/client": "^4.16.2",
+    "@supabase/supabase-js": "^2.45.1",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@types/react-router-dom": "^5.3.3",
     "axios": "^1.4.0",
+    "bcrypt": "^5.1.1",
     "bcryptjs": "^2.4.3",
     "express": "^4.18.2",
     "ioredis": "^5.3.3",
-    "jsonwebtoken": "^9.0.0",
-    "nodemailer": "^6.7.5",
-    "prisma": "^4.0.0"
+    "jsonwebtoken": "^9.0.2",
+    "nodemailer": "^6.9.14",
+    "prisma": "^4.0.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.0.0"
   },
   "devDependencies": {
+    "@types/bcrypt": "^5.0.2",
     "@types/bcryptjs": "^2.4.2",
     "@types/express": "^4.17.13",
+    "@types/jest": "^29.5.12",
     "@types/jsonwebtoken": "^9.0.0",
+    "@types/nodemailer": "^6.4.15",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.2.4",
     "ts-node-dev": "^2.0.0",
-    "typescript": "^5.1.6"
+    "typescript": "^5.5.4"
   },
   "author": "SALAVEY13",
   "license": "MIT"
