@@ -1,7 +1,11 @@
 import { Request, Response, NextFunction } from 'express';
 import jwt from 'jsonwebtoken';
 
-export const authenticateToken = (req: Request, res: Response, next: NextFunction) => {
+interface CustomRequest extends Request {
+  user?: any; // Define this more strictly based on your token payload
+}
+
+export const authenticateToken = (req: CustomRequest, res: Response, next: NextFunction) => {
   const token = req.header('Authorization')?.split(' ')[1];
 
   if (!token) {
@@ -10,7 +14,7 @@ export const authenticateToken = (req: Request, res: Response, next: NextFunctio
 
   try {
     const decoded = jwt.verify(token, process.env.JWT_SECRET as string);
-    req.user = decoded as any;
+    req.user = decoded;
     next();
   } catch (error) {
     res.status(400).json({ error: 'Invalid token' });
