@@ -1,13 +1,13 @@
 import { Router } from 'express';
-import { addBook, getBooks, getBookById, updateBook, deleteBook } from '../controllers/bookController';
-import { authenticateAdmin } from '../middleware/authMiddleware';
+import { createBook, getAllBooks, getBookById, updateBook, deleteBook } from '../controllers/bookController';
+import { authenticateToken } from '../middlewares/authMiddleware';
 
 const router = Router();
 
-router.post('/books', authenticateAdmin, addBook);
-router.get('/books', getBooks);
+router.post('/books', authenticateToken, createBook);
+router.get('/books', getAllBooks);
 router.get('/books/:id', getBookById);
-router.put('/books/:id', authenticateAdmin, updateBook);
-router.delete('/books/:id', authenticateAdmin, deleteBook);
+router.put('/books/:id', authenticateToken, updateBook);
+router.delete('/books/:id', authenticateToken, deleteBook);
 
 export default router;
