@@ -1,6 +1,11 @@
 import Redis from 'ioredis';
 
-const redis = new Redis(process.env.REDIS_URL);
+const redisUrl = process.env.REDIS_URL;
+if (!redisUrl) {
+  throw new Error('REDIS_URL environment variable is not set');
+}
+
+const redis = new Redis(redisUrl);
 
 export const cacheSet = async (key: string, value: any, expireSeconds?: number) => {
   await redis.set(key, JSON.stringify(value));
